+++ test/test_kdtree.cpp
@@ -50,14 +50,14 @@ protected:
 		// For generating random numbers
 		random_device rd;
 		mt19937 gen(rd());
-		uniform_real_distribution<double> real_dis(0.0, static_cast<double>(N));
+		uniform_real_distribution<float> real_dis(0.0, static_cast<float>(N));
 
-		data = (double **)::operator new(N * sizeof(double *));
+		data = (float **)::operator new(N * sizeof(float *));
 		for(i = 0; i < N; i++) {
-			::new(data + i) double *;
-			data[i] = (double *)::operator new(d * sizeof(double));
+			::new(data + i) float *;
+			data[i] = (float *)::operator new(d * sizeof(float));
 			for(j = 0; j < d; j++) {
-				::new(data[i] + j) double;
+				::new(data[i] + j) float;
 				data[i][j] = real_dis(gen);
 			}
 		}
@@ -80,26 +80,26 @@ protected:
 
 public:
 	// Some expensive resource shared by all tests.
-	static double ** data;
+	static float ** data;
 	static int N, d;
 };
 
-double ** KDTreeTest::data;
+float ** KDTreeTest::data;
 int KDTreeTest::N;
 int KDTreeTest::d;
 
 TEST_F(KDTreeTest, DISABLED_test1) {
-	KDNode<double> * a, * b;
-	a = ::new KDNode<double>(d);
-	b = ::new KDNode<double>(d);
+	KDNode<float> * a, * b;
+	a = ::new KDNode<float>(d);
+	b = ::new KDNode<float>(d);
 	a->add_data(data[0]);
 	b->add_data(data[1]);
 	EXPECT_LT(0.0, kd_distance(a,b,false));
 }
 
 TEST_F(KDTreeTest, DISABLED_test2) {
-	KDNode<double> * root = nullptr;
-	double d[][2] = {{2.0,3.0},{4.0,3.0},{7.0,9.0}};
+	KDNode<float> * root = nullptr;
+	float d[][2] = {{2.0,3.0},{4.0,3.0},{7.0,9.0}};
 	for(int i = 0; i < 3; i++) {
 		kd_insert(root,d[i],2,0,i,false);
 	}
@@ -107,24 +107,24 @@ TEST_F(KDTreeTest, DISABLED_test2) {
 }
 
 TEST_F(KDTreeTest, DISABLED_test3) {
-	double ** _d;
-	_d = new double*[10000];
+	float ** _d;
+	_d = new float*[10000];
 	for(int i = 0; i < 10000; i++) {
-		_d[i] = new double[2];
-		_d[i][0] = _d[i][1] = static_cast<double>(i);
+		_d[i] = new float[2];
+		_d[i][0] = _d[i][1] = static_cast<float>(i);
 	}
 	EXPECT_EQ(5000,find_median(_d,10000,2,1,true));
 	//	cout << find_median(data,N,d,64) << endl;
 }
 
 TEST_F(KDTreeTest, DISABLED_test4) {
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_balanced_tree(root,data,N,d,0,0,false);
 	kd_travel(root,d,0);
 }
 
 TEST_F(KDTreeTest, DISABLED_test5) {
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_random_tree(root,data,N,d,0,false);
 	kd_travel(root,d,0);
 }
@@ -132,15 +132,15 @@ TEST_F(KDTreeTest, DISABLED_test5) {
 TEST_F(KDTreeTest, test6) {
 	unsigned long int t1, t2, t3;
 	t1 = get_millisecond_time();
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_random_tree(root,data,N,d,0,false);
 	t2 = get_millisecond_time();
-	//	kd_travel<double>(root,d,0);
-	KDNode<double> * query = ::new KDNode<double>(d);
+	//	kd_travel<float>(root,d,0);
+	KDNode<float> * query = ::new KDNode<float>(d);
 	size_t pos = 10;
 	query->add_data(data[pos]);
-	KDNode<double> * result = nullptr;
-	double best_dist = DBL_MAX;
+	KDNode<float> * result = nullptr;
+	float best_dist = DBL_MAX;
 	size_t visited = 0;
 	nn_search(root,query,result,best_dist,d,0,visited,false);
 	t3 = get_millisecond_time();
@@ -153,15 +153,15 @@ TEST_F(KDTreeTest, test6) {
 TEST_F(KDTreeTest, test7) {
 	unsigned long int t1, t2, t3;
 	t1 = get_millisecond_time();
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_balanced_tree(root,data,N,d,0,0,false);
 	t2 = get_millisecond_time();
-	//	kd_travel<double>(root,d,0);
-	KDNode<double> * query = ::new KDNode<double>(d);
+	//	kd_travel<float>(root,d,0);
+	KDNode<float> * query = ::new KDNode<float>(d);
 	size_t pos = 10;
 	query->add_data(data[pos]);
-	KDNode<double> * result = nullptr;
-	double best_dist = DBL_MAX;
+	KDNode<float> * result = nullptr;
+	float best_dist = DBL_MAX;
 	size_t visited = 0;
 	nn_search(root,query,result,best_dist,d,0,visited,false);
 	t3 = get_millisecond_time();
@@ -172,26 +172,26 @@ TEST_F(KDTreeTest, test7) {
 }
 
 TEST_F(KDTreeTest, DISABLED_test8) {
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_balanced_tree(root,data,N,d,0,0,false);
 	size_t pos = 10;
 	size_t best = 0;
-	double best_dist = DBL_MAX;
+	float best_dist = DBL_MAX;
 	linear_search(data,data[pos],best,best_dist,N,d,true);
 	EXPECT_EQ(0.0,best_dist);
 }
 
 TEST_F(KDTreeTest, DISABLED_test9) {
-	KDNode<double> * a, * b;
-	a = ::new KDNode<double>(d);
-	b = ::new KDNode<double>(d);
+	KDNode<float> * a, * b;
+	a = ::new KDNode<float>(d);
+	b = ::new KDNode<float>(d);
 	a->add_data(data[0]);
 	b->add_data(data[1]);
 	unsigned long int t1, t2, t3;
 	t1 = get_millisecond_time();
-	double d1 = kd_distance(a,b,true);
+	float d1 = kd_distance(a,b,true);
 	t2 = get_millisecond_time();
-	double d2 = distance(data[0],data[1],d);
+	float d2 = distance(data[0],data[1],d);
 	t3 = get_millisecond_time();
 	cout << "kd_distance's time:" << t2-t1 << "[ms]" << endl;
 	cout << "distance's time:" << t3-t2 << "[ms]" << endl;
@@ -199,14 +199,14 @@ TEST_F(KDTreeTest, DISABLED_test9) {
 }
 
 TEST_F(KDTreeTest, test10) {
-	KDNode<double> * root = nullptr;
+	KDNode<float> * root = nullptr;
 	make_random_tree(root,data,N,d,0,false);
-	//	kd_travel<double>(root,d,0);
-	KDNode<double> * query = ::new KDNode<double>(d);
+	//	kd_travel<float>(root,d,0);
+	KDNode<float> * query = ::new KDNode<float>(d);
 	size_t pos = 10;
 	query->add_data(data[pos]);
-	KDNode<double> * result = nullptr;
-	double best_dist = DBL_MAX;
+	KDNode<float> * result = nullptr;
+	float best_dist = DBL_MAX;
 	size_t visited = 0;
 	ann_search(root,query,result,best_dist,1.5,d,0,visited,false);
 	cout << "best distances " << best_dist << endl;
